requirements.txt
asgiref==3.5.2
Django==4.0.5
sqlparse==0.4.2
tzdata==2022.1
djangorestframework==3.13.1
Pillow==8.4.0
Markdown==3.3.7
django-filter==21.1
djangorestframework-simplejwt==5.2.0